TIP-842: Move family variant updater to Akeneo\Pim\Structure
